requirements.txt
flet==0.25.0
pycryptodome==3.21.0